[patch] refactor global_parameters context
