+++ lib/ehealth/web/controllers/global_parameter_controller.ex
@@ -3,12 +3,12 @@ defmodule EHealth.Web.GlobalParameterController do
 
   use EHealth.Web, :controller
 
-  alias EHealth.PRM.GlobalParameters
+  alias EHealth.GlobalParameters
 
   action_fallback EHealth.Web.FallbackController
 
   def index(conn, _params) do
-    with global_parameters <- GlobalParameters.list_global_parameters() do
+    with global_parameters <- GlobalParameters.list() do
       render(conn, "index.json", global_parameters: global_parameters)
     end
   end
@@ -16,7 +16,7 @@ defmodule EHealth.Web.GlobalParameterController do
   def create_or_update(%Plug.Conn{req_headers: req_headers} = conn, params) do
     client_id = get_client_id(req_headers)
 
-    with {:ok, global_parameters} <- GlobalParameters.create_or_update_global_parameters(params, client_id) do
+    with {:ok, global_parameters} <- GlobalParameters.create_or_update(params, client_id) do
       render(conn, "index.json", global_parameters: global_parameters)
     end
   end
