+++ priv/prm_repo/seeds.exs
@@ -1,6 +1,6 @@
 # PRM seed data
 
-alias EHealth.PRM.GlobalParameters.Schema, as: GlobalParameter
+alias EHealth.GlobalParameters.GlobalParameter
 alias EHealth.PRMRepo
 
 :ehealth
